[refactor]: 完成选择角色的 sql 查询重构，目前进入游戏还是会无响应，已得到相关 busy sql 信息，等待进一步重构

+++ src/main/java/com/github/mrzhqiang/maplestory/wz/element/data/Vector.java
@@ -77,8 +77,8 @@ public final class Vector {
      * @since copy from java\awt\geom\Point2D.java
      */
     public double distanceSq(Vector pt) {
-        double px = pt.x - this.x;
-        double py = pt.y - this.y;
+        int px = pt.x - this.x;
+        int py = pt.y - this.y;
         return (px * px + py * py);
     }
